- Fix unintended page refresh when updating the map. - Sort Dash app params alphabetically?
